Dom's comments, refactors to avoid duplicate code
