+++ geoh5py/groups/giftools/magnetic_vector_inversion.py
@@ -21,8 +21,7 @@ from __future__ import annotations
 from typing import Any
 from uuid import UUID
 
-from geoh5py.groups.giftools.base import BaseGIFtoolsGroup
-from geoh5py.groups.giftools.parameters import BASE_PARAMETERS, merge_field
+from geoh5py.groups.giftools.base import BASE_PARAMETERS, BaseGIFtoolsGroup, merge_field
 from geoh5py.groups.giftools.potential_field_base import (
     BOUND_MODEL_LOWER_FIELD,
     POTENTIAL_FIELD_PARAMETERS,
