+++ geoh5py/groups/giftools/potential_field_base.py
@@ -21,8 +21,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from geoh5py.groups.giftools.parameters import (
-    BASE_LENGTH_SCALES,
+from geoh5py.groups.giftools.base import (
     BASE_PARAMETERS,
     merge_field,
 )
@@ -82,27 +81,8 @@ POTENTIAL_FIELD_PARAMETERS: dict[str, Any] = {
     },
     # add main=False relative to the universal base.
     "active_model": merge_field(BASE_PARAMETERS["active_model"], main=False),
-    "auto_threshold": {
-        "choiceList": ["Relative error", "Specify value"],
-        "default": "Relative error",
-        "enabled": True,
-        "group": "Wavelet compression",
-        "groupOptional": True,
-        "label": "Threshold mode",
-        "value": "Relative error",
-    },
-    "bound_model_upper": {
-        "association": "Cell",
-        "dataType": "Float",
-        "default": 10.0,
-        "group": "Model parameters",
-        "isValue": True,
-        "label": "Bounds (upper)",
-        "main": False,
-        "parent": "mesh",
-        "property": "",
-        "value": 10.0,
-    },
+    "auto_threshold": BASE_PARAMETERS["auto_threshold"],
+    "bound_model_upper": BASE_PARAMETERS["bound_model_upper"],
     "cell_weight": BASE_PARAMETERS["cell_weight"],
     "default_decay": {
         "default": False,
@@ -137,29 +117,10 @@ POTENTIAL_FIELD_PARAMETERS: dict[str, Any] = {
     },
     "face_weight": BASE_PARAMETERS["face_weight"],
     "initial_model": BASE_PARAMETERS["initial_model"],
-    "inversion_mode": {
-        "choiceList": ["Chifactor", "Single beta"],
-        "default": "Chifactor",
-        "group": "Inversion stopping criteria",
-        "label": "Inversion mode",
-        "value": "Chifactor",
-    },
-    "inversion_par": {
-        "default": 1.0,
-        "group": "Inversion stopping criteria",
-        "label": "Chifactor or beta",
-        "min": 1e-12,
-        "value": 1.0,
-    },
-    "inversion_tolerance": {
-        "default": 5.0e-2,
-        "group": "Inversion stopping criteria",
-        "label": "Tolerance",
-        "max": 0.5,
-        "min": 1e-3,
-        "value": 5.0e-2,
-    },
-    "length_scales": merge_field(BASE_LENGTH_SCALES, is_length=True),
+    "inversion_mode": BASE_PARAMETERS["inversion_mode"],
+    "inversion_par": BASE_PARAMETERS["inversion_par"],
+    "inversion_tolerance": BASE_PARAMETERS["inversion_tolerance"],
+    "length_scales": BASE_PARAMETERS["length_scales"],
     "lp_epsilon": {
         "default": -1234567,
         "enabled": False,
@@ -184,71 +145,15 @@ POTENTIAL_FIELD_PARAMETERS: dict[str, Any] = {
         "label": "ENZ epsilon",
         "value": 1e-5,
     },
-    "matrix_file": {
-        "enabled": False,
-        "fileDescription": ["UBC matrix file"],
-        "fileType": ["mtx"],
-        "label": "Sensitivity file",
-        "optional": True,
-        "value": "",
-    },
-    "mesh": {
-        "default": "",
-        "label": "Mesh",
-        "main": True,
-        "meshType": "{b020a277-90e2-4cd7-84d6-612ee3f25051}",
-        "value": "",
-    },
-    "reference_model": {
-        "association": "Cell",
-        "dataType": "Float",
-        "default": 1.0e-8,
-        "group": "Model parameters",
-        "isValue": True,
-        "label": "Reference model",
-        "main": False,
-        "parent": "mesh",
-        "property": "",
-        "value": 1.0e-8,
-    },
+    "matrix_file": BASE_PARAMETERS["matrix_file"],
+    "mesh": BASE_PARAMETERS["mesh"],
+    "reference_model": BASE_PARAMETERS["reference_model"],
     "results_loaded": BASE_PARAMETERS["results_loaded"],
     "smooth_mod": BASE_PARAMETERS["smooth_mod"],
-    "threshold": {
-        "default": 5.0e-2,
-        "enabled": True,
-        "group": "Wavelet compression",
-        "groupOptional": True,
-        "label": "Threshold value",
-        "value": 5.0e-2,
-    },
-    "topography": {
-        "default": "",
-        "enabled": False,
-        "label": "Topography",
-        "main": True,
-        "meshType": "{f26feba3-aded-494b-b9e9-b2bbcbe298e1}",
-        "optional": True,
-        "value": "",
-    },
+    "threshold": BASE_PARAMETERS["threshold"],
+    "topography": BASE_PARAMETERS["topography"],
     "version": "6",
-    "wavelet": {
-        "choiceList": [
-            "daub1",
-            "daub2",
-            "daub3",
-            "daub4",
-            "daub5",
-            "daub6",
-            "symm4",
-            "symm5",
-            "symm6",
-        ],
-        "enabled": True,
-        "group": "Wavelet compression",
-        "groupOptional": True,
-        "label": "Wavelet type",
-        "value": "daub2",
-    },
+    "wavelet": BASE_PARAMETERS["wavelet"],
     "wavelet_diagnostics": {
         "default": False,
         "group": "Wavelet compression",
@@ -260,13 +165,7 @@ POTENTIAL_FIELD_PARAMETERS: dict[str, Any] = {
         "value": False,
     },
     "working_directory": BASE_PARAMETERS["working_directory"],
-    "xy_localize": {
-        "default": False,
-        "label": "Localize coordinates",
-        "main": True,
-        "tooltip": "Writes files to disk with respect to UBC origin of 3D grid",
-        "value": False,
-    },
+    "xy_localize": BASE_PARAMETERS["xy_localize"],
 }
 
 # base "bound_model_lower" field that can be easily edited for any specific differences,
